Use default region on bucket location 403 error
Before making a call to GetObject, we look up the location (i.e. region)
for the bucket. However, this may fail with a 403 AccessDenied error if
we don't have permissions on the bucket, but only a subset of objects
within the bucket. In order to still access those objects, we try using
the default region. In the case that we do need bucket permissions for
our operation (e.g. if we're doing a PutObject) or we don't actually
have permission to read the objects in the bucket, we'll still get a 403
when we do the actual operation. I don't see the point in erroring out
when we try to get the bucket location - it seems premature.