Add Region config option
Sometimes, AWS credentials might grant access to certain files on s3,
but not to their bucket. This results in a 403 Access Denied error when
trying to look up the bucket location. This change adds a `Region`
option to the client config that allows the caller to hardcode the
region instead of looking it up in s3.

Tested manually.

Note that the `disableS3Encryption` config option probably doesn't work
for consumers of the library since it's lowercased, making it private to
the package. I considered that out of scope for this change though.