Reset reader before writing
It seems like AWS's SDK used to have this behavior, but somewhere it was taken out.
Likely lives in this diff: https://github.com/aws/aws-sdk-go/compare/v1.12.62...v1.19.15
But it's not immediately obvious where it was changed.

This encodes that behavior into our abstraction layer so we are shielded
by this type of bug in the future.

----------

CC @clever-dan-torczynski who encountered this in the past in a worker